Homework1/DockerCompose/DataCollector/requirements.txt
mysql-connector-python
yfinance
confluent-kafka


